--- src/main/java/me/knighthat/interactivedeck/observable/Observable.java
@@ -1,54 +1,0 @@
-/*
- * Copyright (c) 2023. Knight Hat
- * All rights reserved.
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction,
- * including without limitation the rights to use,copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished
- * to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
- * NONINFRINGEMENT.IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF
- * OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-package me.knighthat.interactivedeck.observable;
-
-import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-
-import java.util.HashSet;
-import java.util.Optional;
-import java.util.Set;
-
-public final class Observable<T> {
-
-    @Nullable T value;
-
-    @NotNull Set<Observer<T>> observers = new HashSet<>();
-
-    Observable( @Nullable T value ) {
-        this.value = value;
-    }
-
-    public static <T> @NotNull Observable<T> of( @Nullable T value ) {
-        return new Observable<>( value );
-    }
-
-    public void value( @Nullable T value ) {
-        this.value = value;
-        notifyObservers();
-    }
-
-    public @NotNull Optional<T> value() {
-        return Optional.ofNullable( value );
-    }
-
-    public void observe( @NotNull Observer<T> observer ) {
-        observers.add( observer );
-    }
-
-    private void notifyObservers() {
-        observers.forEach( observer -> observer.update( value ) );
-    }
-}
